requirements.txt
# Core LangChain & LangGraph
langgraph>=0.4.0
langchain>=0.4.0
langchain-deepseek>=0.1.0
langchain-core>=0.4.0
langchain-community>=0.4.0
langchain-chroma>=0.2.0

# Vector Database
chromadb>=0.5.0

# Checkpointing
langgraph-checkpoint>=2.0.0
langgraph-checkpoint-sqlite>=2.0.0

# Web Search
tavily-python>=0.4.0

# Browser Automation
playwright>=1.50.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Utilities
python-dotenv>=1.0.0
tiktoken>=0.7.0
pydantic>=2.0.0
pillow>=10.0.0
aiohttp>=3.9.0

# OpenAI-compatible client (for DeepSeek via OpenAI compat layer)
openai>=1.0.0

# ── Document Processing ────────────────────────────────────────────────────────
python-docx>=1.1.0
pypdf>=4.0.0
pandas>=2.0.0

# ── Web Server ─────────────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sse-starlette>=2.2.0
